pyproject.toml
[project]
name = "react-typescript-django-channel-drf"
version = "0.1.0"
description = ""
authors = [
    {name = "bluebamus",email = "bluebamus@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
